+++ ThirdPartyAdapters/chartboost/chartboost/src/main/java/com/google/ads/mediation/chartboost/ChartboostAdapter.java
@@ -41,7 +41,7 @@ public class ChartboostAdapter implements MediationRewardedVideoAdAdapter,
     /**
      * The current version of the adapter.
      */
-    public static final String ADAPTER_VERSION_NAME = "7.3.0.0";
+    public static final String ADAPTER_VERSION_NAME = "7.3.1.0";
 
     /**
      * Key to obtain App ID, required for initializing Chartboost SDK.