+++ ThirdPartyAdapters/facebook/facebook/src/main/java/com/google/ads/mediation/facebook/FacebookAdapter.java
@@ -17,9 +17,9 @@ package com.google.ads.mediation.facebook;
 import android.content.Context;
 import android.content.pm.PackageManager;
 import android.content.res.Resources;
-import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.Keep;
 import android.text.TextUtils;
@@ -33,7 +33,6 @@ import android.widget.ImageView;
 import android.widget.RelativeLayout;
 
 import com.facebook.ads.Ad;
-import com.facebook.ads.AdChoicesView;
 import com.facebook.ads.AdError;
 import com.facebook.ads.AdListener;
 import com.facebook.ads.AdOptionsView;
@@ -45,8 +44,8 @@ import com.facebook.ads.InterstitialAdListener;
 import com.facebook.ads.MediaView;
 import com.facebook.ads.MediaViewListener;
 import com.facebook.ads.NativeAd;
+import com.facebook.ads.NativeAdLayout;
 import com.facebook.ads.NativeAdListener;
-import com.facebook.ads.NativeAdViewAttributes;
 import com.facebook.ads.RewardedVideoAd;
 import com.facebook.ads.RewardedVideoAdListener;
 import com.google.android.gms.ads.AdRequest;
@@ -71,6 +70,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Mediation adapter for Facebook Audience Network.
@@ -87,6 +87,8 @@ public final class FacebookAdapter
 
     private static final int MAX_STAR_RATING = 5;
     private static final String TAG = "FacebookAdapter";
+
+    private static AtomicInteger gmsVersion = new AtomicInteger(0);
 
     private MediationBannerListener mBannerListener;
     private MediationInterstitialListener mInterstitialListener;
@@ -121,12 +123,6 @@ public final class FacebookAdapter
      * sent to the Google Mobile Ads SDK.
      */
     private boolean mIsImpressionRecorded;
-
-    /**
-     * Flag to determine whether or not to make AdChoices icon for native ads expandable.
-     * {@code true} by default.
-     */
-    private boolean mIsAdChoicesIconExpandable = true;
 
     /**
      * A Facebook {@link MediaView} used to show native ad media content.
@@ -179,7 +175,7 @@ public final class FacebookAdapter
                                 MediationAdRequest adRequest,
                                 Bundle mediationExtras) {
         mBannerListener = listener;
-        if(!mIsSdkInitialized.getAndSet(true)) {
+        if (!mIsSdkInitialized.getAndSet(true)) {
             AudienceNetworkAds.initialize(context);
         }
         if (!isValidRequestParameters(context, serverParameters)) {
@@ -202,7 +198,6 @@ public final class FacebookAdapter
             mBannerListener.onAdFailedToLoad(this, AdRequest.ERROR_CODE_NO_FILL);
             return;
         }
-        AdSettings.setMediationService("ADMOB_" + FacebookAdapter.getGMSVersionCode(context));
         mAdView = new AdView(context, placementId, facebookAdSize);
         mAdView.setAdListener(new BannerListener());
         buildAdRequest(adRequest);
@@ -211,7 +206,20 @@ public final class FacebookAdapter
         mWrappedAdView = new RelativeLayout(context);
         mAdView.setLayoutParams(adViewLayoutParams);
         mWrappedAdView.addView(mAdView);
-        mAdView.loadAd();
+        if (gmsVersion.get() == 0) {
+            GMSVersionUpdateTask gmsVersionUpdateTask = new GMSVersionUpdateTask(context);
+             gmsVersionUpdateTask.Task(new GMSVersionUpdateTask.OnGMSUpdateListener() {
+                @Override
+                public void OnGMSVersionUpdate() {
+                    AdSettings.setMediationService("ADMOB_" + gmsVersion);
+                    mAdView.loadAd();
+                }
+            });
+            gmsVersionUpdateTask.execute();
+        } else {
+            AdSettings.setMediationService("ADMOB_" + gmsVersion);
+            mAdView.loadAd();
+        }
     }
 
     @Override
@@ -228,7 +236,7 @@ public final class FacebookAdapter
                                       MediationAdRequest adRequest,
                                       Bundle mediationExtras) {
         mInterstitialListener = listener;
-        if(!mIsSdkInitialized.getAndSet(true)) {
+        if (!mIsSdkInitialized.getAndSet(true)) {
             AudienceNetworkAds.initialize(context);
         }
         if (!isValidRequestParameters(context, serverParameters)) {
@@ -238,11 +246,23 @@ public final class FacebookAdapter
 
         String placementId = serverParameters.getString(PLACEMENT_PARAMETER);
 
-        AdSettings.setMediationService("ADMOB_" + FacebookAdapter.getGMSVersionCode(context));
         mInterstitialAd = new InterstitialAd(context, placementId);
         mInterstitialAd.setAdListener(new InterstitialListener());
         buildAdRequest(adRequest);
-        mInterstitialAd.loadAd();
+        if (gmsVersion.get() == 0) {
+            GMSVersionUpdateTask gmsVersionUpdateTask = new GMSVersionUpdateTask(context);
+             gmsVersionUpdateTask.Task(new GMSVersionUpdateTask.OnGMSUpdateListener() {
+                @Override
+                public void OnGMSVersionUpdate() {
+                    AdSettings.setMediationService("ADMOB_" + gmsVersion);
+                    mInterstitialAd.loadAd();
+                }
+            });
+            gmsVersionUpdateTask.execute();
+        } else {
+            AdSettings.setMediationService("ADMOB_" + gmsVersion);
+            mInterstitialAd.loadAd();
+        }
     }
 
     @Override
@@ -262,7 +282,7 @@ public final class FacebookAdapter
                            Bundle serverParameters,
                            Bundle networkExtras) {
         mContext = context;
-        if(!mIsSdkInitialized.getAndSet(true)) {
+        if (!mIsSdkInitialized.getAndSet(true)) {
             AudienceNetworkAds.initialize(context);
         }
         mRewardedListener = mediationRewardedVideoAdListener;
@@ -289,8 +309,20 @@ public final class FacebookAdapter
             mRewardedListener.onAdLoaded(this);
         } else {
             buildAdRequest(mediationAdRequest);
-            AdSettings.setMediationService("ADMOB_" + FacebookAdapter.getGMSVersionCode(mContext));
-            mRewardedVideoAd.loadAd(true);
+            if (gmsVersion.get() == 0) {
+                GMSVersionUpdateTask gmsVersionUpdateTask = new GMSVersionUpdateTask(mContext);
+                 gmsVersionUpdateTask.Task(new GMSVersionUpdateTask.OnGMSUpdateListener() {
+                    @Override
+                    public void OnGMSVersionUpdate() {
+                        AdSettings.setMediationService("ADMOB_" + gmsVersion);
+                        mRewardedVideoAd.loadAd(true);
+                    }
+                });
+                gmsVersionUpdateTask.execute();
+            } else {
+                AdSettings.setMediationService("ADMOB_" + gmsVersion);
+                mRewardedVideoAd.loadAd(true);
+            }
         }
     }
 
@@ -328,7 +360,7 @@ public final class FacebookAdapter
                                 NativeMediationAdRequest mediationAdRequest,
                                 Bundle mediationExtras) {
         mNativeListener = listener;
-        if(!mIsSdkInitialized.getAndSet(true)) {
+        if (!mIsSdkInitialized.getAndSet(true)) {
             AudienceNetworkAds.initialize(context);
         }
         if (!isValidRequestParameters(context, serverParameters)) {
@@ -347,19 +379,25 @@ public final class FacebookAdapter
 
         String placementId = serverParameters.getString(PLACEMENT_PARAMETER);
 
-        // Get the optional extras if set by the publisher.
-        if (mediationExtras != null) {
-            mIsAdChoicesIconExpandable = mediationExtras.getBoolean(
-                    FacebookExtrasBundleBuilder.KEY_EXPANDABLE_ICON, true);
-        }
-
         mMediaView = new MediaView(context);
 
-        AdSettings.setMediationService("ADMOB_" + FacebookAdapter.getGMSVersionCode(context));
         mNativeAd = new NativeAd(context, placementId);
         mNativeAd.setAdListener(new NativeListener(mNativeAd, mediationAdRequest));
         buildAdRequest(mediationAdRequest);
-        mNativeAd.loadAd();
+        if (gmsVersion.get() == 0) {
+            GMSVersionUpdateTask gmsVersionUpdateTask = new GMSVersionUpdateTask(context);
+             gmsVersionUpdateTask.Task(new GMSVersionUpdateTask.OnGMSUpdateListener() {
+                @Override
+                public void OnGMSVersionUpdate() {
+                    AdSettings.setMediationService("ADMOB_" + gmsVersion);
+                    mNativeAd.loadAd();
+                }
+            });
+            gmsVersionUpdateTask.execute();
+        } else {
+            AdSettings.setMediationService("ADMOB_" + gmsVersion);
+            mNativeAd.loadAd();
+        }
     }
     //endregion
 
@@ -644,7 +682,7 @@ public final class FacebookAdapter
 
         @Override
         public void onMediaDownloaded(Ad ad) {
-            Log.d(TAG,"onMediaDownloaded");
+            Log.d(TAG, "onMediaDownloaded");
         }
     }
 
@@ -676,11 +714,38 @@ public final class FacebookAdapter
         return Math.round(pixel / displayMetrics.density);
     }
 
-    private static int getGMSVersionCode(Context context) {
-        try {
-            return context.getPackageManager().getPackageInfo("com.google.android.gms", 0 ).versionCode;
-        } catch (PackageManager.NameNotFoundException e) {
-            return 0;
+    public static class GMSVersionUpdateTask extends AsyncTask<Void, Void, Integer> {
+
+        public Context context;
+        GMSVersionUpdateTask(Context context) {
+            this.context = context;
+        }
+
+        public interface OnGMSUpdateListener {
+            void OnGMSVersionUpdate();
+        }
+
+        private OnGMSUpdateListener listener;
+
+        private void Task(OnGMSUpdateListener listener) {
+            this.listener = listener;
+        }
+
+        @Override
+        protected Integer doInBackground(Void... voids) {
+            try {
+                return context.getPackageManager().getPackageInfo("com.google.android.gms", 0).versionCode;
+            } catch (PackageManager.NameNotFoundException e) {
+                return 0;
+            }
+        }
+
+        @Override
+        protected void onPostExecute(Integer result) {
+            gmsVersion.set(result);
+            if (this.listener != null) {
+                listener.OnGMSVersionUpdate();
+            }
         }
     }
 
@@ -823,17 +888,20 @@ public final class FacebookAdapter
                                Map<String, View> nonClickableAssetViews) {
 
             ViewGroup adView = (ViewGroup) view;
-
             // Find the overlay view in the given ad view. The overlay view will always be the
             // top most view in the hierarchy.
             View overlayView = adView.getChildAt(adView.getChildCount() - 1);
             if (overlayView instanceof FrameLayout) {
-                // Create and add Facebook's AdChoicesView to the overlay view.
-                AdOptionsView adOptionsView =
-                        new AdOptionsView(view.getContext(), mNativeAd, null);
+                NativeAdLayout nativeAdLayout = new NativeAdLayout(view.getContext());
+                ((FrameLayout) overlayView).addView(nativeAdLayout);
+
+                // Create and add Facebook's AdOptions to the overlay view.
+                AdOptionsView adOptionsView = new AdOptionsView(view.getContext(), mNativeAd,
+                    nativeAdLayout);
+
                 ((ViewGroup) overlayView).addView(adOptionsView);
                 // We know that the overlay view is a FrameLayout, so we get the FrameLayout's
-                // LayoutParams from the AdChoicesView.
+                // LayoutParams from the AdOptionsView.
                 FrameLayout.LayoutParams params =
                         (FrameLayout.LayoutParams) adOptionsView.getLayoutParams();
                 if (mNativeAdOptions != null) {
@@ -857,11 +925,8 @@ public final class FacebookAdapter
                 }
                 adView.requestLayout();
             } else {
-
-                AdOptionsView adOptionsView =
-                        new AdOptionsView(view.getContext(), mNativeAd, null);
+                AdOptionsView adOptionsView = new AdOptionsView(view.getContext(), mNativeAd, null);
                 this.setAdChoicesContent(adOptionsView);
-
             }
 
             // Facebook does its own impression tracking.
@@ -876,8 +941,8 @@ public final class FacebookAdapter
                 assetViews.add(clickableAssets.getValue());
 
                 if (clickableAssets.getKey().equals(NativeAppInstallAd.ASSET_ICON) ||
-                        clickableAssets.getKey().equals(UnifiedNativeAdAssetNames.ASSET_ICON)){
-                    iconview = (ImageView)clickableAssets.getValue();
+                        clickableAssets.getKey().equals(UnifiedNativeAdAssetNames.ASSET_ICON)) {
+                    iconview = (ImageView) clickableAssets.getValue();
                 }
 
             }
@@ -911,35 +976,6 @@ public final class FacebookAdapter
                 return null;
             }
             return (MAX_STAR_RATING * rating.getValue()) / rating.getScale();
-        }
-    }
-
-    /**
-     * The {@link FacebookExtrasBundleBuilder} class is used to create a network extras bundle that
-     * can be passed to the adapter to make network specific customizations.
-     */
-    public static class FacebookExtrasBundleBuilder {
-
-        /**
-         * Key to add and obtain {@link #mIsExpandableIcon}.
-         */
-        private static final String KEY_EXPANDABLE_ICON = "expandable_icon";
-
-        /**
-         * Whether or not ad choices icon for native ads is expandable.
-         */
-        private boolean mIsExpandableIcon;
-
-        public FacebookExtrasBundleBuilder setNativeAdChoicesIconExpandable(
-                boolean isExpandableIcon) {
-            this.mIsExpandableIcon = isExpandableIcon;
-            return FacebookExtrasBundleBuilder.this;
-        }
-
-        public Bundle build() {
-            Bundle bundle = new Bundle();
-            bundle.putBoolean(KEY_EXPANDABLE_ICON, mIsExpandableIcon);
-            return bundle;
         }
     }