+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -1,5 +1,8 @@
 # InMobi Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 7.2.2.0
+- Verified compatibility with InMobi SDK version 7.2.2.
+
 ## 7.2.1.0
 - Verified compatibility with InMobi SDK version 7.2.1.